onpier_commandor/requirements.txt
hvac[parser]
pydantic
jira
kubernetes
pandas
loguru
tqdm
typer-cli
diagrams
requests
GitPython
PyGitHub
httpx
joblib
python-dotenv
typer[all]